Add --save and --if-dirty to install
