+++ datalad/distribution/create.py
@@ -258,7 +258,7 @@ class Create(Interface):
         if dataset is not None and dataset.path != tbds.path:
             # we created a dataset in another dataset
             # -> make submodule
-            from .install import _install_subds_inplace
+            from datalad.distribution.utils import _install_subds_inplace
             subdsrelpath = relpath(realpath(tbds.path), realpath(dataset.path))  # realpath OK
             _install_subds_inplace(ds=dataset, path=tbds.path,
                                    relativepath=subdsrelpath)
